CR TLSbo86683: Fix fec_dcache_range args
fec_dcache_flush_range() and fec_dcache_inv_range() are being passed
start and end addresses, but the arm calls expect address and length.

The patch TLSbo86683 that this fix is based on called
fec_dcache_inv_range with a start address and an end address -4, but a
similar change was not made for fec_dcache_flush_range.  It is not clear
whether the end address should be inclusive or exclusive of the last
word of data. For now, the end address points just past the end of the
data.

http://www.bitshrine.org/gpp/linux-2.6.19.2-mx-fix_fec_dcache_range_args.patch