CR ENGR00028231: Remove workaround for FIR tx SDMA script issue
Description: Earlier SDMA script (before SS09) had problems with
transmit packets if the length in bytes given to it was not a multiple
of 4. For this reason, the workaround was to set the length of the
packet on the transmit side to the immediate next multiple of 4. This
workaround can now be removed as the new SDMA script takes care of this.

http://www.bitshrine.org/gpp/linux-2.6.19.2-mx-remove_fir_tx_sdma_workaround.patch