This patch adds SDHC functionality to the MX31 Rel3 (2.6.19.2) kernel. It is a direct port of a patch posted to lkml by Philip Langdale.  Note that Philip's original patch expects that the mmc high-speed patch has already been applied, which is not the case for the MX31 kernel.
Upstream patch information:
  URL:     http://lkml.org/lkml/2007/1/4/90
  Date:    Thu, 04 Jan 2007 06:57:32 -0800
  From:    Philip Langdale <>
  Subject: [PATCH 2.6.19] mmc: Add support for SDHC cards (Take 5)
------------------------------------------------------------------------
Thanks to the generous donation of an SDHC card by John Gilmore, and the
surprisingly enlightened decision by the SD Card Association to publish
useful specs, I've been able to bash out support for SDHC. The changes
are not too profound:

i) Add a card flag indicating the card uses block level addressing and
check it in the block driver. As we never took advantage of byte-level
addressing, this simply involves skipping the block -> byte translation
when sending commands.

ii) The layout of the CSD is changed - a set of fields are discarded to
make space for a larger C_SIZE. We did not reference any of the
discarded fields except those related to the C_SIZE.

iii) Read and write timeouts are fixed values and not calculated from
CSD values.

iv) Before invoking SEND_APP_OP_COND, we must invoke the new
SEND_IF_COND to inform the card we support SDHC.

I've done some basic read and write tests and everything seems to work
fine but one should obviously use caution in case it eats your data.

Signed-off-by: Philipl Langdale <philipl@overt.org>

http://www.bitshrine.org/gpp/linux-2.6.19.2-mx-mmc_sdhc.patch