Bugzilla 563: MX31 register swizzle work-around
A problem has been discovered with MX31-TO2 silicon where a general
register is loaded with an address constant from a PC-relative location.
Upon going indirect through this register, the address value contained
in the register will have been rotated right by 8 bits, causing a memory
fault. At present the cause of the rotation is not known.

This patch implements a work around to this problem, although why it
works is still unknown.

http://www.bitshrine.org/gpp/linux-2.6.19.2-mx31_TO2_reg_swizzle_workaround.patch