MX27Litekit: Add basic display support
The patches did not contain the board specific information to initialize
the display.  This patch will add support for the Sharp VGA display from
LogicPD.  We would eventually like to add support for other displays as well.

This patch is based on work by the user fszczerba on the LogicPD forums.

Signed-off-by: Justin Waters <justin.waters@timesys.com>