MX27LITE_GPIO: Fix GPIO conflicts between USB Host 2 and SPI 1
USB Host 2 and SPI 1 only conflict with one pin.  If the SPI bus is not using
the slave select 2 (SS2) pin (which the Atlas chip does not), then you can use
both the SPI 1 and USB Host 2 interfaces simultaneously.  This patch disables
the MSB of the slave select if Host 2 is enabled.  This will limit SPI1 to
only 4 slaves instead of 8.

Signed-off-by: Justin Waters <justin.waters@timesys.com>