Update pin configuration for USB Host 2 on the Lite Kit.
The USB Host 2 on the MX27 Litekit uses the USB_POWER and USB_OC_B pins for
power and overcurrent detect, respectively.  In order to make them function
properly, we need to assign those pins to the Host 2 driver and change the
H2PM field of the USB_CTRL register on the i.MX27.

Signed-off-by: Justin Waters <justin.waters@timesys.com>