RTC_MXC: Handle shared interrupts properly
The MXC RTC driver interrupt handler never checked to make sure it was handling
the correct IRQ.  This was causing it to oops whenever CONFIG_DEBUG_SHIRQ was
enabled in the kernel.  This fixes that problem by checking the IRQ status
registers on the RTC module whenever the IRQ occurs.

Signed-off-by: Justin Waters <justin.waters@timesys.com>