MX27 DPTC: Remove Vestigal Code
The Dynamic Processor Temperature Control Module is significantly different on
the MX27 as compared to the MX31.  It seems that some MX31 code slipped into
the MX27 devices file, but essentially did nothing except try to register a
device that didn't have a driver.  I removed this code.

I also fixed an issue with the Kconfig where you could select the DPTC module
without the power management chip enabled, which causes a compilation error.

Signed-off-by: Justin Waters <justin.waters@timesys.com>