Clean up MX27 CRM Register Definitions
The MX27 crm_regs.h file was a complete mess, obviously hastily ported from
the MX31 version, which was also a mess.  I cleaned it up quite a bit,
although more work still needs to be done such as fixing the naming
conventions.

This is build tested on the MX27 Lite Kit.  It may break MX27 builds.

Signed-off-by: Justin Waters <justin.waters@timesys.com>