Remove GFP_DMA flag from mx2fb driver
The GFP_DMA flag is unneccesary when using dma_alloc_coherent, and may
possibly break the code if an initramfs is used.  There is no real reason
why it should be limited to the first 16 MB of memory.  I am removing it
so that it will work with the initramfs system.

Signed-off-by: Justin Waters <justin.waters@timesys.com>