Cleanup mxc touchscreen driver
The mxc touchscreen driver, based on the Freescale Atlas Chip and i.MX chips,
is a mess.  This is my attempt to clean it up quite a bit, and I think it's
a definite improvement.  There are still issues on large displays near the
higher X coordinates, but I believe this is a hardware issue.

Signed-off-by: Justin Waters <justin.waters@timesys.com>