s3c2443-irq: Bug fix for accessing to the EXTINT0-register
There is a HW-bug in the S3C2443: When the register of the
external interrupts (EXTINT0) is read, the returned register
value is swapped. For this reason, it is required to swap
the read bytes before writing into the register. The bug
fix is coming from the SMDK of Samsung.

Signed-off-by: Luis Galdos <luis.galdos@digi.com>