config ccwi9p9215: Added to support new wi9p9215 module
Added this file to support the Wi9p9215 module.  This file is
based on mach-cc9p9215js.c.  The differences are that it calls
ns9xxx_add_device_ccWi9p9215_wifi to initialize the piper wifi
driver, and calls mach_ccwi9p9215js_init_machine to initialize
the module and carrier board.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>