wi9p9215 piper: Added field in RF structure for channel change time
The mac80211 library requires us to tell it how many micro seconds
it will take for us to change channels.  I added an int field in
the RF structure to hold this information.  The value is set by
the RF code and used by the pipermain code which talks to the
mac80211 library.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>