cc9m2443-pm: Added the wakeup-GPIOs
For having an easier handling of the wakeup-GPIOs,
the code includes a small driver for registering
and configuring the GPIO/IRQ. The wakeup can be
enabled over the sysfs-attribute:

    /sys/devices/platform/cc9m2443-wio.232/power/wakeup

The above attribute enables/disables the wakeup
of the GPIO number 232.

The included GPIOs are 226 and 232. They can be
selected over:

    System type > S3C2443 Machines > Enable GPIO ...

Signed-off-by: Luis Galdos <luis.galdos@digi.com>