w9p9215 :Changed platform name from wi9p to w9p
I found that the convention for wireless platform names is
to use just a "w" instead of "wi".  For example, the
wireless version of the cc9c is the ccw9c.  Also, there
were already some provisions for a w9p9215js platform
in some of the existing files.  So, I changed the name
of the new platform from ccwi9p9215js to ccw9p9215js.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>