cc9m2443-pm: Removed the internal driver for the wakeup-GPIOs
In the commit[1] an internal driver for the control
of the wakeup-GPIOs was added. The problem with this
implemtation was the missing reference to the GPIO-
structure passed to the GPIOLIB (see [2]).

This commit removes the internal driver, so that the
extended GPIOLIB is now responsible for enabling or
disabling the wakeup-GPIOs.

[1] da4f2f9e28c66256eae5fbbd57fa219340b80897
[2] arch/arm/mach-s3c2443/gpio.c:s3c2443_gpios

Signed-off-by: Luis Galdos <luis.galdos@digi.com>