ns9xxx-eth: remove BUG_ON assertions
The driver was crashing when using a wrong phyid from user-space
applicatins (for example snmpd), and as a result the whole module
was hung.

With the commit the driver returns with -EINVAL instead of hanging in
case of incorrect phyid. This solves the problem of some user-space
applications killing the system.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>