smsc911x: fixes duplex management and modifies ioctl
This commit fixes the management of the duplex mode when forcing
the link mode.
The mii_check_media returns  when the media is forced, therefore,
the duplex mode was not correctly updated in the mac registers.
The commit also modifies the ioctl to call generic_mii_icotl instead
of implementing the access to the mii registers directly in the
funciton. This allows to check the duplex change and call the new
function created smsc911x_phy_update_duplex, in that situation.

Signed-off-by: Pedro Perez de Heredia <pperez@pperez-desktop.(none)>