s3c2443_pcmcia: fix PM support on PCMCIA driver
Fixes incorrect Power Management handling on the S3C2443 SoC.
The device wake up functionality from the card detect line has
been removed from the driver because not all IRQ lines are
wake up capable. The code should be generalized and extended in
future to support that, if supported by the irq.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>