ACK problem:  Fixed problem with missing ACKs
Tracked down problem in transmitter routine which would cause
us to stall.  Found that sometimes when we transmit a packet
and expect an ACK, that we get neither the ACK or a timeout
interrupt.  I don't fully understand why that happens, but I
suspect that the H/W receives an ACK and so cancels the
timeout interrupt, and then can lose the ACK if a receive
overflow occurs.  Fixed the problem by starting a S/W timer
on every transmit.  The timer fires in 250 mills, unless we
receive an ACK or a timeout interrupt.  The timer routine
processes the event as if we received an ACK.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>