ns921x-serial: Added RS-485 transceiver control support.
NS921x serial ports have the posibility to use RTS line
for RS-485 transceiver control. The patch adds this
configuration option to the serial ports selection and
the driver enables the RTSEN flag in the Wrapper
Configuration register of the UART.
This configuration only requests three GPIOs (RX/TX/RTS).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>