ns9360: initialize correctly video interface
If the selected video interface is the VGA, on NS9360 platforms, the
processor uses the external clock to derive the pixel clock.

This means that the video driver needs to initialize the clock register
to enable the external video clock. To do so the gpio must be properly
configured as special function.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>