pm s3c2443: hacked to support PCMCIA PM functionality on the CC9M2443JS board
When PCMCIA is enabled, there is a problem when wakeing up from
sleep. this is becasue the register EXTINT(0,1,2) should be saved
before going to sleep and restored after the wake up interrupt.

On the S3C2443 there is a problem to read that register being
necessary to use a special access sequence.
That sequence is unknown at the moment and, for that reason, we
just reconfigure here the card detect gpio line.
This hack should be removed in the near future and some additional
code should be added here to save/restore some of the external
interrupt registers.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>