piper: fix DUP packets under ping
Fixes problem with dup packets when pinging from host to target
and viceversa.
The get_rate function was not returning the correct rate to
retransmit when the last packet was not acked by the AP.
The retransmition counter(s) are now updated inside the interrupt
handler (under the timeout section).
The get_rate function has been also slightly modified.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>