ccw9p9215 wifi: Fixed BRS setting for 802.11a
Fixed bug in routine that sets the BRS register.  The mac80211 library
passes us this information in a different format for 802.11a channels.
Had to update the routine to deal with that.

Also changed the routine that sets the transmitter power to store
the new setting in the local context rather than having this done
by multiple helper routines.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>