ccw9p9215 wifi: Fixed infinite retry problem
Fixed a bug that could cause us to retry a packet endlessly under
some circumstances.  The function to determine the transmit rate
for a packet could be overridden by the calibration transmit
rate during calibration.  The problem was that in this case the
retry counter was not being updated, so the transmitter would
never stop retrying a frame while calibration was taking place.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>