ccw9p9215 wifi: Updated data structure for new calibration data format
Updated the calibration data structure definition from the 1.1 format
to the 1.2 format.  The change is that the max airoha index field for
each curve becomes the maximum ADC value for each cureve.

Updated prototype for the getOfdmBrs function (which now requires
a channel number), and added a prototype for set_tracking_constant.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>