ccw9p9215 wifi: Fixed bug in NVRAM routine that prevented reload
The calibration routines have to convert the version 1.1 format
calibration data to version 1.2 format.  When they do this, the
checksum becomes invalid.  This caused the checksum test to fail
if the driver was shut down and restarted.  Fixed this by having
the routine adjust the version number and recalculate the checksum
when it converts 1.1 data to 1.2 format.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>