ccw9p9215 wifi: Fixed hang when no calibration data
Fixed a problem that caused the calibration state machine to hang
when it was unloaded if no calibration data was present.  The problem
was that it would keep retrying to read the calibration data.  Since
the data never changes, the read would always fail and the state
machine would never progress onto the next step.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>