ccw9p9215 wifi: Updated ieee80211_ioctl_girange to pass txpower info
The ieee80211_ioctl_girange function did not fill in the transmit
power information in the iw_range structure passed to it.  This
would cause the command "iwlist wlan0 txpower" to fail with the
message "unknown transmit-power information".  I updated the function
to fill in the appropriate fields in the structure.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>