ccw9p9215 wifi: Fine tuned duty cycling algorithm
Fixed problem in routine that predicts when next beacon occurs.

Changed routine that calculates when to duty cycle to do so based
upon the time between when the current beacon is received and when
the next one is suppose to arrive.  This should make the routine
be better able to adjust when the apparent time beacons arrive
changes due to H/W problems.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>