ccw9p9215 wifi: Handled spike suppression problem
Piper has a bug in it which causes noise spikes when transmitting at
1 and 2 Mbps unless a special spike suppression routine is run when
piper is powered up.  The problem is that this routine takes an
average of 30ms, and I have measured it taking as long as 300 ms.
This is not acceptable for the power save routines which run on a
100 ms cycle.  I added code to work around this by changing the
transmit rate for 1 and 2 Mbps packets to be 5.5 Mbps when running
in power save mode.  This allows us to power up without doing the
noise suppression routine.  The noise suppression routine will be
executed when we exit power save mode.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>