ccw9p9215 wifi: Limited transmit when tx queue filling and Schaffner fixes
Changed the PS routines so that they will not try to transmit PS-Poll
packets when the transmit queue starts filling up.

Changed the power up code as advised by Mike Schaffner.

Added a guard to the tx dequeue algorithm so that we will not go past
the beginning of the queue.  Also added code into the tx_done routine
to try prevent us from processing a NULL packet if we get a duplicate
ACK.  I have seen APs send us duplicate ACKS, so this is a possibility.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>