aes: Free DMA allocation on setkey method
The setkey method allocates a DMA area for a future
call to the crypto method. When the crypto operation
finishes the area is freed. However, a call may be
made to the setkey method that isn't followed by a
crypto operation, a second call to setkey leaves us
with an unfreed area, which may lead to a NOMEM error
(see Vantive 31266).
This patch fixes this problem by evaluating the dma_addr_t
involved and freeing the are if it was already allocated,
before allocating a new one for the same address.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>