pwm: remove overflow check and fix Kconfig
An overflow function was checking values exceeding
2^32, but doing a string comparision. This lead to
overflows on any number starting with a character
bigger than '5'.
The overflow check is not needed because the 'sscanf'
function is used to parse values as unsigned long, and
this is already truncating values bigger than 2^32 to
that limit.

The dependencies on the Kconfig for the PWM support didn't
let you enable SPI support in the kernel. The correct
dependency should check for the platform SPI master.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>