ccw9p9215js: switch off wifi status led at startup
This commit ensures that the wifi status led is off (by switching it)
even when the wireless driver is not selected.
This shoudld be done by the bootloader that should leave that led on
when booting the OS.
Another fix will go to the bootloader.
It also removes the code added by a previous commit that tried to fix
this issue (and that didnt solve it).

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>