ccwi9p9215 wifi: Added code to power cycle the transmitter
Added code to power down the transmitter and then power it back up
again when changing channels.  There is a possiblity that we might
be transmitting a frame when we change channels.  If that happens
it could screw up the tranceiver.  It's better to power down the
transmitter when changing channels.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>