ccw9p9215 wifi: Fixed problem where changing channel set wrong power
I fixed a problem where the routines to change the channel and to
initialise the RF hardware would set the power level using an
uncalibrated value.  Added a hook into the calibration routine which
can be used by the channel changing code to read the correct power
index to use.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>