ccw9p9215 wifi: Corrected problems wtih suspend and Cisco switch
SA found that the suspend/resume function didn't work with Cisco switches.
The AP would refuse to associate with us for a while after we resumed.
Fixed this by sending a NULL data frame with power management set
before suspending, and then another with the power management bit clear
when we resume.

Also updated the suspend/resume function to sychronize with the
power save function, and to also use the power save implementations of
the functions to shut down and power up the wifi hardware.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>