piper: Remove setting PS flag in frame control for each frame
This commit removes the code that sets the Power Management flag
in the frame control section of the header for each frame.
Only the PSPoll and the nullDF will set this flag in the header.

This has been changed because the previous implementation was
causing problems with the Cisco Aironet 1200 when the protected
and the PM flags were both set in a frame.

This should be revisited in future and implemented accord to the
specification.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>