uncompress: s3c putc wrongly assumes serial is enabled
The putc() function wrongly assumes that the serial port is
enabled by the bootloader. However, the bootloader may not
even initialize the serial port if the silent console feature
is activated. This leads to and endless loop checking for the
FIFO empty flag (which never happens if the port is not enabled
for transmitting).

This patch checks in UCON register whether the TXMODE field
is enabled. If it is not, the putc() function does not print
anyhing.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>