cc9m2443js: fixed ARM machine ID number
Originally, there were two registered machine IDs:
  1663      ConnectCore 9M 2443                     cc9m2443
  1815      ConnectCore 9M 2443 on JumpStart        cc9m2443js

Due to an error both in the kernel and in U-Boot, the first one
(1663) was being used for the ConnectCore 9M 2443 on JumpStart
board. Since the code error happened both in the boot loader and
in the kernel, U-Boot was able to boot the kernel without
problems.

In order to correct the code but maintain compatibility with old
boot loaders and kernels, the machines IDs have been changed in
the ARM developer's list and the code fixed both in the kernel
of DEL-5.2 master branch and U-Boot DUB RevF3 master branch, so
that the IDs look like this:
  1663          ConnectCore 9M 2443 on JumpStart        cc9m2443js
  1815          ConnectCore 9M 2443                     cc9m2443

So, starting from DEL-5.2 and U-Boot DUB RevF3, the code is fixed.
The change is however transparent to old boot loaders and kernels
because the ID passed by the bootloader and expected by Linux is
still 1663 for the ConnectCore 9M 2443 on JumpStart board.
In general terms, we can state that this is a sort of aesthetic
code change more than a functional one.

The reason for the change, is also due to the ability of U-Boot
to use a user given machine ID. For example, when trying to boot
a cc9m2443js kernel on a ccw9m2443 module, if you use the,
theoritical correct value 1815 for the cc9m2443js (according to
the old list), the kernel won't boot. Now the list shows the
correct ID.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>