ns9360-serial: enable console UART during setup
If a UART port different than the one used by U-Boot console
was selected as Linux kernel console (by changing the 'console'
parameter in U-Boot), no messages were printed and the system
didn't boot.
Enabling the UART to be used as console, during the console setup
function (instead of enabling it for each transfer), solves the
problem.
I tried to debug the kernel to see why the enabling on the write
function did not do the job, but the debugger doesn't work well
at this early part of the kernel.

This patch solves vantive 31387.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>