kconfig: clean compiler warnings
Dirty hack to avoid compiler warnings such as:

warning: ignoring return value of ‘fgets’, declared with attribute
warn_unused_result
warning: ignoring return value of ‘asprintf’, declared with attribute
warn_unused_result

This happens with gcc-4.3, but not with previous versions of gcc. Our
base distribution for Digi Embedded Linux (kubuntu) uses gcc-4.3 as
host compiler so for now we need this hack to avoid customers getting
warnings on kernel compilation.

This must be removed when upstream fixes the warnings or we update our
base distribution to a newer version of gcc.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>