wi9p: Experimental fix for wifi transmit rate algorithm
This change makes the PID algorithm specify the retry rates so that the
primary rate is only tried a few times, and then we will try successively
lower rates until we either successfully transmit the packet or run out
of retries.  Originally the algorithm would only retry with the primary
rate.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>