wi9p: Fixed problem with Minstrel Tx rate algorithm.
Found and fixed the problem that prevented the Minstrel TX rate
algorithm from working.  It turns out that Mintrel wants the list
of retry statistics to be terminated by setting the index of the
entry after the entry that succeeds to -1.  It would have been
nice if this had been mentioned somewhere in the documentation.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>