wi9p: Reversed the antenna connectors
Originally I set up the antenna map so that the connector closest to the
antenna labeled primary would carry the primary signal.  This was so that
users wouldn't have to cross the wires.  However, NET+OS set up the map
so that users would have to cross the wires.  Since the Linux product is
not officially released yet, we decided to change it to match NET+OS so
that we are consistent across the product lines.  Eventually, the labeling
for the antennae on the carrier board will either be corrected or removed.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>