wi9p: Changed name of new PID based wireless tx rate algorithm
I originally named the new PID algorithm digipid.  This was not a
good name because it incorporated the company name.  I changed
the code to rename the new algorithm multirate-pid.

I also fixed some minor bugs in the new code that prevented
accurate statistics from being recorded by the debug file system.
I also put in a guard in the routine that gets the rate that if,
somehow, the maxrates is 1, we will only set the first rate count
to 1 instead of 2, which is what we were doing unconditionally.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>