ns9xxx-eth: support Eth activity LED with a gpio
The ConnectME module has the Ethernet activity LED directly
connected to a GPIO.
This patch adds the support for using a GPIO for emulating the
Eth RX/TX activity LED.

The toggling of the GPIO is done by the timeout function of
a kernel timer: if there has been activity in the last timer
run, then the GPIO is toggled. If not, a counter is incremented
and the timer is retriggered. If the counter reaches the switch
off limit (there has not been activity for some time), the LED
is switched off. This prevents it from staying ON due to an odd
number of toggles.

Any RX/TX interrupt will trigger the activity flag and the timer
in case it is not already running.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>