ccwi9m2443-piper: fix PM problem when waking up from suspend
This commit fixes some problems with the PM support and the piper
driver, on the CCWi9M2443 platform. The issue is that the PM support
of the s3c2443 doenst reconfigure the irq lines as they were before
suspending the system. Its also necessary to reconfigure the CS line
to ensure we can access the piper chip after wake up.
This should be fixed in future in the PM support of the s3c2443 chip
and then it can be removed from here.
The current workaround invokes a platform callback when waking up the
unit. This allows the different platforms to make any initialization
that could be necessary when comming from suspend.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>