ccw9p9215: Updated MAC Assist firmware to version 1.16
Updated the MAC Assist firmware to version 1.16.  According to Mike
Schaffner, this update corrects a problem in processing OFDM frames
where the latency would be longer than the inter-frame space
setting.  This problem only occured with certain rates and frame
lengths.  Also, the value of the minimum IFS counter was incremented
from 1 to 2.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>