s3c2443 irq: EXTINTx registers wrongly read
As stated in Samsung document "S3C2443 Guide to extra GPIO",
Some GPIO registers and EXTINT0, EXTINT1, EXTINT2, can be directly
written but cannot be read in the standard way.
For reading, a nibbles swapping is needed.
Erroneously, the swapping was being done only for EXTINT0, but not
for EXTINT1 and EXTINT2. As a consequence, a first write to this
register was OK (the register is 0 after reset), but a second write
was not done ok (as an OR is done with the previously read value).

The problem was discovered when enabling PCMCIA card detect interrupt
and USB Overload protection interrupt, which are both configured in
EXTINT1. If only one of the interfaces was enabled, it worked fine,
but if both were enabled, the second write was causing the misconfiguration
of the previous interrupt (for example PCMCIA card detect was reset
to LOW LEVEL ACTIVE) and hence the interrupt was triggered all the time,
causing a system hang (Vantive 32058).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>