s3c2443 irq: EXTINTx registers wrongly read (2nd)
The practical tests in U-Boot have shown the following when writing
and reading EXTINT registers:
1. Writing seems to be ok
2. For reading, the following happens:
  EXTINT0:
	1. Read values are given in big endian (at nibble level):
	   A read value of 0x00001234 is really 0x43210000.
	2. The read value does not return the LSB:
	   After writing 0xffffffff value read is 0x7fffffff.
  EXTINT1/2:
	1. Read values are given in big endian (at nibble level)
	and bits at nibble level suffer a 1 bit rotation to the left:
	   After writing 0x12488822 value read is 0x44111842

This patch does the necessary transformation to the nibbles so that
EXTINT1/2 can be read properly, and thus correctly written when a
bitfield must be written and is masked with the previously read value.

irqs.h: add comments to better locate interrupts numbers.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>