ccw9p: Cleaned up duty cycling shut down code
The code to shut down duty cycling did not shut down the transmission
process cleanly.  It was possible for it to leave the driver in a
bad state.  I rewrote that portion of the code to handle the piper
shutdown just like the regular duty cycling routine does.  This
includes sending a null-data frame with the power management bit set,
shutting off the transmit queues, and waiting for the pending transmit
to complete before powering down piper.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>