wi9p: Added scanning guard to duty cycling
Found that mac80211 turns or leaves duty cycling on while wpa_supplicant
is attempting to scan and/or reassociate.  This interferes with the
process and can cause wpa_supplicant to spin endlessly trying to
reassociate.  Added code into the mac80211 entry points to set a timer
when mac80211 does something that looks like it is scanning or
reassociation (changing channels for example). Duty cycling will now
shut down for a while when one of these events occurs.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>