wi9p: Disabled duty cycle throttling
Disabled the code that throttles duty cycling when we receive association
type events.  This code seems like it might reduce our performance due to
interactions with the mac80211 library and/or the wireless AP.  The driver
seems to run better without it.

Also disabled the debugging printfs in preparations for release.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>