wi9p: Changed set channel routine so it doesn't turn receiver on
Previously the set channel routine always enabled the receiver before
it completed.  This was because it had to disable it at the beginning
of the routine.  However, the set channel routine is used by the
power save duty cycling code, and this code needs to have the receiver
off when the routine returns.  I changed the code so that there are
two entry points.  The new entry point allows power save to use the
routine without enabling the receiver.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>