wi9p: Changed power up routine to use a new version of set channel
The original version of set channel had a side effect of turning the
receiver on.  This caused problems because the rest of piper was not
yet configured when the routine was called.  Added a new set channel
routine which does not enable the receiver, and changed the power up
routine to use it.

Also added more debug code which is conditionally compiled when
WANT_DEBUG is defined.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>