wi9p: Set Airoha register 0xC to 0x2b
This change was actually already in the code.  This modification just
makes it more explicit.  We check the platform type and hardware
revision, and store it.  Whenever the Airoha 7230 code resets the
transciever or sets the channel it uses the stored information to
correctly set the value of register 0xC, which will be 0x2b for
2 GHz channels.

Pedro plans on updating the code for the Wi9M.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>