scsi: allow SCSI_WAIT_SCAN to be disabled in embedded systems
I know scsi_wait_scan is too little and is always built as a module,
but:

* This allows it to be disabled in cases that modules are enabled but
all configuration is built-in. It will not surprise users that a module
has been built.

* It still does not surprise users that have disabled it by accident,
because they only have a choice if they've picked up EMBEDDED.

* This change allows it to be built-in. It uses late_initcall, which
will put it ahead of other (SCSI) drivers.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@holoscopio.com>
Signed-off-by: Javier Viguera <javier.viguera@digi.com>