ENGR00120807 V4L2 capture:Change the mechanism for CSI->MEM capture
As CSI->MEM channel keeps on writing the buffer which is set to
ready latestly and raising up end of frame interrupts, the current
v4l2 capture realization mechanism is not appropriate for this channel.
This patch will update the idmac buffer to a dummy buffer whenever
there is no buffers queued by the user.

Signed-off-by: Liu Ying <b17645@freescale.com>
Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>