ENGR00121003 Fix IRAM enabling for USB Gadget driver
In order to add IRAM to the USB Gadget driver, there are 2 kconfig
options to enable. One kconfig option was for USB host and the other
was for Gadget. The current SW required both kconfig to be
enable which is confusing. Also need to add iram_alloc function in
the Gadget driver.

Signed-off-by: Dinh Nguyen <Dinh.Nguyen@freescale.com>
Signed-off-by: Alejandro Gonzalez <alex.gonzalez@digi.com>