wi9p wifi: Limited maximum transmit rate on selected 802.11a channels
This change was requested by Nacho.  The Wi9p hardware does not pass
when some 802.11a channels are used at 54M or 48M.  To deal with this
the driver will now limit those channels to a maximum of 36M or 48M
as necessary.  Due to the limits of the Linux wireless infrastructure,
we will still advertise that we support 54M on all channels, but when
the driver receives a request to set a rate it cannot support it will
simply set the highest rate that it can.  Also, it will report that
it sent the frame at the higher rate.

Signed-off-by: Charles Gordon <charles.gordon@digi.com>